+++ lib/foreman_ansible_core/runner/command_creator.rb
